rename "views" to "routes" for more clarity, flatten directory structure
